refactor: Sub Component 폴더 구조 변경
