+++ front/src/components/Create/OptionPropertyForm/style.ts
