+++ front/src/components/Detail/style.ts
