+++ front/src/components/index.ts
@@ -1,19 +1,19 @@
 export { CategoryList } from './CategoryList';
-export { CategoryListForm } from './CategoryListForm';
+export { CategoryListForm } from './Create/CategoryListForm';
 export { Counter } from './Counter';
 export { DefaultInfo } from './DefaultInfo';
 export { Header } from './Header';
 export { Option } from './Option';
 export { OptionProperty } from './OptionProperty';
 export { Product } from './Product';
-export { ProductDetail } from './ProductDetail';
+export { Detail } from './Detail';
 export { ProductList } from './ProductList';
 export { Tag } from './Tag';
-export { TagForm } from './TagForm';
+export { TagForm } from './Create/TagForm';
 export { TagList } from './TagList';
-export { TagListForm } from './TagListForm';
+export { TagListForm } from './Create/TagListForm';
 export { Cart } from './Cart';
-export { Card } from './Card';
+export { Card } from './Cart/Card';
 export { Create } from './Create';
-export { OptionForm } from './OptionForm';
-export { OptionPropertyForm } from './OptionPropertyForm';
+export { OptionForm } from './Create/OptionForm';
+export { OptionPropertyForm } from './Create/OptionPropertyForm';
